refactor: 💡 rm stale integration test for real test server
