+++ package.json
@@ -10,13 +10,11 @@
     "generate": "prisma generate",
     "start": "next start",
     "lint": "next lint && eslint --fix --ext .js,.jsx,.ts,.tsx .",
-    "test": "npm run test:supertest",
-    "test:integration": "TEST_TYPE=integration jest --config=jest.integration.config.ts",
+    "test": "jest",
+    "test:integration": "TEST_TYPE=integration jest --config=jest.supertest.config.ts",
     "test:unit": "jest --testPathIgnorePatterns=integration",
     "test:debug": "DEBUG_TESTS=true jest --verbose --silent=false",
-    "test:integration:debug": "TEST_TYPE=integration DEBUG_TESTS=true jest --config=jest.integration.config.ts --verbose --silent=false",
-    "test:supertest": "TEST_TYPE=integration jest --config=jest.supertest.config.ts",
-    "test:supertest:debug": "TEST_TYPE=integration DEBUG_TESTS=true jest --config=jest.supertest.config.ts --verbose --silent=false",
+    "test:integration:debug": "TEST_TYPE=integration DEBUG_TESTS=true jest --config=jest.supertest.config.ts --verbose --silent=false",
     "test:unit:debug": "DEBUG_TESTS=true jest --testPathIgnorePatterns=integration --verbose --silent=false",
     "check-format": "prettier --ignore-path .gitignore --list-different \"**/*.+(js|jsx|ts|tsx|json|css)\"",
     "format": "prettier --ignore-path .gitignore --write \"**/*.+(js|jsx|ts|tsx|json|css)\"",
