+++ src/tests/integration/test_cases/00_api_health_check.test.ts
@@ -1,6 +1,6 @@
 // Info: (20240701 - Shirley) Supertest-based API health check integration tests
 // Info: (20240701 - Shirley) Focus on API functionality without database dependencies
-import { createTestClient, TestClient } from '@/tests/integration/supertest/test_client';
+import { createTestClient, TestClient } from '@/tests/integration/test_client';
 import statusInfoHandler from '@/pages/api/v2/status_info';
 // import { ITeam } from '@/interfaces/team';
