--- src/tests/integration/test_cases/02_team_management.test.ts
@@ -1,857 +1,0 @@
-import { DefaultValue } from '@/constants/default_value';
-import { ApiClient } from '@/tests/integration/api_client';
-import { SharedTestServer } from '@/tests/integration/shared_server';
-
-/**
- * Integration Test - Team Management & Setup (Ticket #2)
- *
- * Follows Test Case Time Estimation planning:
- * - Test Case 2.1: Team Creation API Testing (2h)
- * - Test Case 2.2: Team Member Invitation Testing (3h)
- * - Test Case 2.3: Team Member Management Testing (3h)
- *
- * Testing Philosophy:
- * - Tests team creation with valid team data, authentication validation, and permission checks
- * - Tests member invitation API with email validation, permission checks, and invitation tracking
- * - Tests member listing, role updates, and member removal with proper permission validation
- * - Validates both success and failure scenarios
- * - Ensures proper error handling and status messages
- * - Tests role hierarchy restrictions and business rules
- */
-describe('Integration Test - Team Management & Setup (Ticket #2)', () => {
-  const testEmails = DefaultValue.EMAIL_LOGIN.EMAIL;
-  const defaultCode = DefaultValue.EMAIL_LOGIN.CODE;
-  let sharedServer: SharedTestServer;
-
-  // 使用共享測試服務器
-  beforeAll(async () => {
-    sharedServer = await SharedTestServer.getInstance();
-    // 只在 debug 模式下啟用詳細 API 輸出
-    if (process.env.DEBUG_TESTS === 'true') {
-      process.env.DEBUG_API = 'true';
-    }
-  }, 30000); // 減少timeout因為服務器已經在運行
-
-  // 不需要afterAll因為全局清理會處理
-
-  // ========================================
-  // Test Case 2.1: Team Creation API Testing
-  // ========================================
-
-  describe('Test Case 2.1: Team Creation API Testing', () => {
-    describe('POST /api/v2/team - Team Creation Success and Failure Scenarios', () => {
-      let apiClient: ApiClient;
-      let userEmail: string;
-
-      beforeAll(() => {
-        apiClient = new ApiClient(sharedServer.getBaseUrl());
-        [userEmail] = testEmails; // Use first test email
-      });
-
-      beforeEach(async () => {
-        // Clean test state before each test
-        await sharedServer.cleanTestState();
-        apiClient.clearSession();
-
-        // Authenticate using HTTP API calls (same as 01 integration test)
-        try {
-          // First, create email login record via HTTP API
-          const emailResponse = await apiClient.get(`/api/v2/email/${userEmail}/one_time_password`);
-          // eslint-disable-next-line no-console
-          console.log('📧 Email authentication response:', JSON.stringify(emailResponse, null, 2));
-          expect(emailResponse).toBeDefined();
-          expect(emailResponse.success).toBe(true);
-
-          // Then authenticate with verification code via HTTP API
-          const loginResponse = await apiClient.post(
-            `/api/v2/email/${userEmail}/one_time_password`,
-            {
-              code: defaultCode,
-            }
-          );
-          // eslint-disable-next-line no-console
-          console.log('🔐 Login response:', JSON.stringify(loginResponse, null, 2));
-          expect(loginResponse).toBeDefined();
-          expect(loginResponse.success).toBe(true);
-
-          // Verify session state with status_info after login
-          const statusResponse = await apiClient.get('/api/v2/status_info');
-          // eslint-disable-next-line no-console
-          console.log('📋 Status info response:', JSON.stringify(statusResponse, null, 2));
-          expect(statusResponse).toBeDefined();
-          expect(statusResponse.success).toBe(true);
-
-          // Verify we got user information and not unauthorized
-          if (statusResponse.success && statusResponse.payload) {
-            const statusData = statusResponse.payload as { user?: { email?: string } };
-            expect(statusData.user).toBeDefined();
-            expect(statusData.user?.email).toBe(userEmail);
-            // eslint-disable-next-line no-console
-            console.log('✅ Successfully retrieved user info for:', statusData.user?.email);
-          }
-        } catch (error) {
-          // eslint-disable-next-line no-console
-          console.log('❌ Authentication or status check failed:', error);
-          expect(error).toBeDefined();
-        }
-      });
-
-      it('should successfully create team with valid data', async () => {
-        const teamData = {
-          name: `Test Team ${Date.now()}`,
-          description: 'Integration test team for API validation',
-        };
-
-        try {
-          const createTeamResponse = await apiClient.post('/api/v2/team', teamData);
-          expect(createTeamResponse).toBeDefined();
-          expect(typeof createTeamResponse.success).toBe('boolean');
-
-          if (createTeamResponse.success) {
-            const teamPayload = createTeamResponse.payload as { id: number; name: string };
-            expect(teamPayload).toBeDefined();
-            expect(teamPayload.id).toBeDefined();
-            expect(teamPayload.name).toBe(teamData.name);
-          }
-        } catch (error) {
-          expect(error).toBeDefined();
-        }
-      });
-
-      it('should create team with minimum required data (name only)', async () => {
-        const teamData = {
-          name: `Minimal Team ${Date.now()}`,
-        };
-
-        try {
-          const createTeamResponse = await apiClient.post('/api/v2/team', teamData);
-          expect(createTeamResponse).toBeDefined();
-          expect(typeof createTeamResponse.success).toBe('boolean');
-
-          if (createTeamResponse.success) {
-            const teamPayload = createTeamResponse.payload as { id: number; name: string };
-            expect(teamPayload.name).toBe(teamData.name);
-          }
-        } catch (error) {
-          expect(error).toBeDefined();
-        }
-      });
-
-      it('should fail team creation with invalid data (empty name)', async () => {
-        const invalidTeamData = {
-          name: '',
-          description: 'Team with empty name',
-        };
-
-        try {
-          const createTeamResponse = await apiClient.post('/api/v2/team', invalidTeamData);
-
-          // Should either fail with validation error or return failure response
-          if (createTeamResponse.success === false) {
-            expect(createTeamResponse.success).toBe(false);
-          }
-        } catch (error) {
-          // Expect validation error for empty name
-          expect(error).toBeDefined();
-        }
-      });
-
-      it('should fail team creation with missing required fields', async () => {
-        const invalidTeamData = {}; // Missing name field
-
-        try {
-          const createTeamResponse = await apiClient.post('/api/v2/team', invalidTeamData);
-
-          if (createTeamResponse.success === false) {
-            expect(createTeamResponse.success).toBe(false);
-          }
-        } catch (error) {
-          expect(error).toBeDefined();
-        }
-      });
-
-      it('should verify team icon generation and default settings initialization', async () => {
-        const teamData = {
-          name: `Icon Test Team ${Date.now()}`,
-          description: 'Testing team icon and default settings',
-        };
-
-        try {
-          const createTeamResponse = await apiClient.post('/api/v2/team', teamData);
-          // eslint-disable-next-line no-console
-          console.log('🏢 Create team response:', JSON.stringify(createTeamResponse, null, 2));
-
-          expect(createTeamResponse).toBeDefined();
-          expect(typeof createTeamResponse.success).toBe('boolean');
-
-          if (createTeamResponse.success) {
-            const teamPayload = createTeamResponse.payload as {
-              id: number;
-              name: string;
-              imageId?: string;
-            };
-
-            expect(teamPayload.id).toBeDefined();
-            expect(teamPayload.name).toBe(teamData.name);
-
-            // Team should have some form of icon/image identification
-            if ('imageId' in teamPayload) {
-              expect(teamPayload.imageId).toBeDefined();
-            }
-
-            // eslint-disable-next-line no-console
-            console.log('✅ Successfully created team with ID:', teamPayload.id);
-          } else {
-            // If team creation fails, log the error for debugging
-            // eslint-disable-next-line no-console
-            console.log('❌ Team creation failed:', createTeamResponse);
-          }
-        } catch (error) {
-          // eslint-disable-next-line no-console
-          console.log('❌ Team creation error:', error);
-          expect(error).toBeDefined();
-        }
-      });
-
-      // Based on Integration Test Plan 2.1: Additional failure test cases
-      it('should handle unauthorized access (no authentication)', async () => {
-        // Clear session to test unauthorized access
-        apiClient.clearSession();
-
-        const teamData = {
-          name: `Unauthorized Test Team ${Date.now()}`,
-          description: 'Testing unauthorized access',
-        };
-
-        try {
-          const createTeamResponse = await apiClient.post('/api/v2/team', teamData);
-
-          // Should either fail with 401 or return failure response
-          if (createTeamResponse.success === false) {
-            expect(createTeamResponse.success).toBe(false);
-            // Expect 401ISF0000 "Unauthorized access" based on test plan
-            expect(createTeamResponse.code).toContain('401');
-          }
-        } catch (error) {
-          // Expect unauthorized error
-          expect(error).toBeDefined();
-          // eslint-disable-next-line no-console
-          console.log('✅ Correctly rejected unauthorized team creation');
-        }
-      });
-
-      it('should validate input parameters and return appropriate errors', async () => {
-        // Test various invalid input scenarios from Integration Test Plan
-
-        const invalidTestCases = [
-          {
-            name: 'Empty team data',
-            data: {},
-            expectedErrorPattern: '422', // Invalid input parameter
-          },
-          {
-            name: 'Missing name field',
-            data: { description: 'Team without name' },
-            expectedErrorPattern: '422', // Invalid input parameter
-          },
-          {
-            name: 'Invalid data format',
-            data: { name: 123, description: 'Invalid name type' },
-            expectedErrorPattern: '422', // Invalid input parameter
-          },
-        ];
-
-        // eslint-disable-next-line no-restricted-syntax
-        for (const testCase of invalidTestCases) {
-          try {
-            // eslint-disable-next-line no-console
-            console.log(`🧪 Testing: ${testCase.name}`);
-            // eslint-disable-next-line no-await-in-loop
-            const createTeamResponse = await apiClient.post('/api/v2/team', testCase.data);
-
-            // Should return failure response with appropriate error code
-            if (createTeamResponse.success === false) {
-              expect(createTeamResponse.success).toBe(false);
-              expect(createTeamResponse.code).toContain(testCase.expectedErrorPattern);
-              // eslint-disable-next-line no-console
-              console.log(
-                `✅ ${testCase.name}: Correctly returned error ${createTeamResponse.code}`
-              );
-            }
-          } catch (error) {
-            // API might throw exception for invalid data
-            expect(error).toBeDefined();
-            // eslint-disable-next-line no-console
-            console.log(`✅ ${testCase.name}: Correctly threw exception`);
-          }
-        }
-      });
-    });
-  });
-
-  // ========================================
-  // Test Case 2.2: Team Member Invitation Testing
-  // ========================================
-
-  describe('Test Case 2.2: Team Member Invitation Testing', () => {
-    describe('PUT /api/v2/team/{teamId}/member - Team Member Invitation Workflow', () => {
-      let apiClient: ApiClient;
-      let userEmail: string;
-      let testTeamId: number;
-
-      beforeAll(() => {
-        apiClient = new ApiClient();
-        [userEmail] = testEmails;
-      });
-
-      afterAll(() => {
-        apiClient.clearSession();
-      });
-
-      beforeEach(async () => {
-        // Authenticate using HTTP API calls (same as 01 integration test)
-        try {
-          // First, create email login record via HTTP API
-          const emailResponse = await apiClient.get(`/api/v2/email/${userEmail}/one_time_password`);
-          // eslint-disable-next-line no-console
-          console.log('📧 Email authentication response:', JSON.stringify(emailResponse, null, 2));
-          expect(emailResponse).toBeDefined();
-          expect(emailResponse.success).toBe(true);
-
-          // Then authenticate with verification code via HTTP API
-          const loginResponse = await apiClient.post(
-            `/api/v2/email/${userEmail}/one_time_password`,
-            {
-              code: defaultCode,
-            }
-          );
-          // eslint-disable-next-line no-console
-          console.log('🔐 Login response:', JSON.stringify(loginResponse, null, 2));
-          expect(loginResponse).toBeDefined();
-          expect(loginResponse.success).toBe(true);
-
-          // Verify session state with status_info after login
-          const statusResponse = await apiClient.get('/api/v2/status_info');
-          // eslint-disable-next-line no-console
-          console.log('📋 Status info response:', JSON.stringify(statusResponse, null, 2));
-          expect(statusResponse).toBeDefined();
-          expect(statusResponse.success).toBe(true);
-
-          // Verify we got user information and not unauthorized
-          if (statusResponse.success && statusResponse.payload) {
-            const statusData = statusResponse.payload as { user?: { email?: string } };
-            expect(statusData.user).toBeDefined();
-            expect(statusData.user?.email).toBe(userEmail);
-            // eslint-disable-next-line no-console
-            console.log('✅ Successfully retrieved user info for:', statusData.user?.email);
-          }
-
-          // Create a test team for invitation testing
-          const teamData = {
-            name: `Invitation Test Team ${Date.now()}`,
-            description: 'Team for testing member invitations',
-          };
-
-          const createTeamResponse = await apiClient.post('/api/v2/team', teamData);
-          if (createTeamResponse.success) {
-            const teamPayload = createTeamResponse.payload as { id: number };
-            testTeamId = teamPayload.id;
-            // eslint-disable-next-line no-console
-            console.log('✅ Created test team with ID:', testTeamId);
-          }
-        } catch (error) {
-          // eslint-disable-next-line no-console
-          console.log('❌ Authentication, status check, or team creation failed:', error);
-          expect(error).toBeDefined();
-        }
-      });
-
-      it('should successfully invite team members with valid emails', async () => {
-        if (!testTeamId) {
-          // eslint-disable-next-line no-console
-          // eslint-disable-next-line no-console
-          console.log('Skipping test: No test team created');
-          return;
-        }
-
-        const invitationData = {
-          emails: [testEmails[1], testEmails[2]], // Use emails from DefaultValue
-        };
-
-        try {
-          const inviteResponse = await apiClient.put(
-            `/api/v2/team/${testTeamId}/member`,
-            invitationData
-          );
-
-          expect(inviteResponse).toBeDefined();
-          expect(typeof inviteResponse.success).toBe('boolean');
-
-          if (inviteResponse.success) {
-            // Verify invitation response structure
-            const invitationResult = inviteResponse.payload;
-            expect(invitationResult).toBeDefined();
-          }
-        } catch (error) {
-          expect(error).toBeDefined();
-        }
-      });
-
-      it('should fail invitation with invalid email formats', async () => {
-        if (!testTeamId) {
-          // eslint-disable-next-line no-console
-          // eslint-disable-next-line no-console
-          console.log('Skipping test: No test team created');
-          return;
-        }
-
-        const invalidInvitationData = {
-          emails: ['invalid-email', 'another-invalid-email@'],
-        };
-
-        try {
-          const inviteResponse = await apiClient.put(
-            `/api/v2/team/${testTeamId}/member`,
-            invalidInvitationData
-          );
-
-          // Should fail with validation error
-          if (inviteResponse.success === false) {
-            expect(inviteResponse.success).toBe(false);
-          }
-        } catch (error) {
-          // Expect validation error for invalid emails
-          expect(error).toBeDefined();
-        }
-      });
-
-      it('should handle empty email list in invitation', async () => {
-        if (!testTeamId) {
-          // eslint-disable-next-line no-console
-          // eslint-disable-next-line no-console
-          console.log('Skipping test: No test team created');
-          return;
-        }
-
-        const emptyInvitationData = {
-          emails: [],
-        };
-
-        try {
-          const inviteResponse = await apiClient.put(
-            `/api/v2/team/${testTeamId}/member`,
-            emptyInvitationData
-          );
-
-          if (inviteResponse.success === false) {
-            expect(inviteResponse.success).toBe(false);
-          }
-        } catch (error) {
-          expect(error).toBeDefined();
-        }
-      });
-
-      it('should validate invitation permission requirements', async () => {
-        if (!testTeamId) {
-          // eslint-disable-next-line no-console
-          // eslint-disable-next-line no-console
-          console.log('Skipping test: No test team created');
-          return;
-        }
-
-        // Test that invitation requires proper permissions (INVITE_MEMBER)
-        const invitationData = {
-          emails: [testEmails[1]], // Use email from DefaultValue
-        };
-
-        try {
-          const inviteResponse = await apiClient.put(
-            `/api/v2/team/${testTeamId}/member`,
-            invitationData
-          );
-
-          expect(inviteResponse).toBeDefined();
-          expect(typeof inviteResponse.success).toBe('boolean');
-
-          // Permission validation should be handled by the API
-          // Either success with proper permissions or failure without them
-        } catch (error) {
-          expect(error).toBeDefined();
-        }
-      });
-
-      it('should track invitation status and workflow', async () => {
-        if (!testTeamId) {
-          // eslint-disable-next-line no-console
-          // eslint-disable-next-line no-console
-          console.log('Skipping test: No test team created');
-          return;
-        }
-
-        const invitationData = {
-          emails: [testEmails[2]], // Use email from DefaultValue
-        };
-
-        try {
-          const inviteResponse = await apiClient.put(
-            `/api/v2/team/${testTeamId}/member`,
-            invitationData
-          );
-
-          if (inviteResponse.success) {
-            const invitationResult = inviteResponse.payload;
-
-            // Verify invitation tracking data structure
-            expect(invitationResult).toBeDefined();
-
-            // Should contain information about invitation status
-            // This could include invitation IDs, status, or other tracking info
-          }
-        } catch (error) {
-          expect(error).toBeDefined();
-        }
-      });
-    });
-  });
-
-  // ========================================
-  // Test Case 2.3: Team Member Management Testing
-  // ========================================
-
-  describe('Test Case 2.3: Team Member Management Testing', () => {
-    describe('Team Member Role Updates and Removal', () => {
-      let apiClient: ApiClient;
-      let userEmail: string;
-      let testTeamId: number;
-
-      beforeAll(() => {
-        apiClient = new ApiClient();
-        [userEmail] = testEmails;
-      });
-
-      afterAll(() => {
-        apiClient.clearSession();
-      });
-
-      beforeEach(async () => {
-        // Authenticate using HTTP API calls (same as 01 integration test)
-        try {
-          // First, create email login record via HTTP API
-          const emailResponse = await apiClient.get(`/api/v2/email/${userEmail}/one_time_password`);
-          // eslint-disable-next-line no-console
-          console.log('📧 Email authentication response:', JSON.stringify(emailResponse, null, 2));
-          expect(emailResponse).toBeDefined();
-          expect(emailResponse.success).toBe(true);
-
-          // Then authenticate with verification code via HTTP API
-          const loginResponse = await apiClient.post(
-            `/api/v2/email/${userEmail}/one_time_password`,
-            {
-              code: defaultCode,
-            }
-          );
-          // eslint-disable-next-line no-console
-          console.log('🔐 Login response:', JSON.stringify(loginResponse, null, 2));
-          expect(loginResponse).toBeDefined();
-          expect(loginResponse.success).toBe(true);
-
-          // Verify session state with status_info after login
-          const statusResponse = await apiClient.get('/api/v2/status_info');
-          // eslint-disable-next-line no-console
-          console.log('📋 Status info response:', JSON.stringify(statusResponse, null, 2));
-          expect(statusResponse).toBeDefined();
-          expect(statusResponse.success).toBe(true);
-
-          // Verify we got user information and not unauthorized
-          if (statusResponse.success && statusResponse.payload) {
-            const statusData = statusResponse.payload as { user?: { email?: string } };
-            expect(statusData.user).toBeDefined();
-            expect(statusData.user?.email).toBe(userEmail);
-            // eslint-disable-next-line no-console
-            console.log('✅ Successfully retrieved user info for:', statusData.user?.email);
-          }
-
-          // Create a test team for member management testing
-          const teamData = {
-            name: `Member Management Test Team ${Date.now()}`,
-            description: 'Team for testing member management operations',
-          };
-
-          const createTeamResponse = await apiClient.post('/api/v2/team', teamData);
-          if (createTeamResponse.success) {
-            const teamPayload = createTeamResponse.payload as { id: number };
-            testTeamId = teamPayload.id;
-            // eslint-disable-next-line no-console
-            console.log('✅ Created test team with ID:', testTeamId);
-          }
-        } catch (error) {
-          // eslint-disable-next-line no-console
-          console.log('❌ Authentication, status check, or team creation failed:', error);
-          expect(error).toBeDefined();
-        }
-      });
-
-      it('should list team members with proper pagination', async () => {
-        if (!testTeamId) {
-          // eslint-disable-next-line no-console
-          // eslint-disable-next-line no-console
-          console.log('Skipping test: No test team created');
-          return;
-        }
-
-        try {
-          // Test basic member listing
-          let membersResponse = await apiClient.get(`/api/v2/team/${testTeamId}/member`);
-          expect(membersResponse).toBeDefined();
-          expect(typeof membersResponse.success).toBe('boolean');
-
-          if (membersResponse.success) {
-            const membersData = membersResponse.payload;
-            expect(membersData).toBeDefined();
-
-            // Should be paginated data structure
-            if (typeof membersData === 'object' && membersData !== null) {
-              // Might contain data, totalCount, page, pageSize, etc.
-            }
-          }
-
-          // Test with pagination parameters
-          membersResponse = await apiClient.get(
-            `/api/v2/team/${testTeamId}/member?page=1&pageSize=10`
-          );
-          expect(membersResponse).toBeDefined();
-        } catch (error) {
-          expect(error).toBeDefined();
-        }
-      });
-
-      it('should validate LIST_MEMBER_BY_TEAM_ID permission', async () => {
-        if (!testTeamId) {
-          // eslint-disable-next-line no-console
-          // eslint-disable-next-line no-console
-          console.log('Skipping test: No test team created');
-          return;
-        }
-
-        try {
-          const membersResponse = await apiClient.get(`/api/v2/team/${testTeamId}/member`);
-
-          // Should either succeed with proper permissions or fail without them
-          expect(membersResponse).toBeDefined();
-          expect(typeof membersResponse.success).toBe('boolean');
-
-          // Permission validation is handled by the API
-          // The test validates the response structure is correct
-        } catch (error) {
-          expect(error).toBeDefined();
-        }
-      });
-
-      it('should update member roles with proper permissions', async () => {
-        if (!testTeamId) {
-          // eslint-disable-next-line no-console
-          // eslint-disable-next-line no-console
-          console.log('Skipping test: No test team created');
-          return;
-        }
-
-        // First get team members to find a member to update
-        try {
-          const membersResponse = await apiClient.get(`/api/v2/team/${testTeamId}/member`);
-
-          if (membersResponse.success) {
-            const membersData = membersResponse.payload;
-
-            // If there are members, try to update role
-            if (membersData && typeof membersData === 'object') {
-              // This would require existing member data
-              // For now, test the API endpoint structure
-              const testMemberId = 'test-member-id';
-
-              const roleUpdateData = {
-                role: 'EDITOR', // or other valid role
-              };
-
-              try {
-                const updateResponse = await apiClient.put(
-                  `/api/v2/team/${testTeamId}/member/${testMemberId}`,
-                  roleUpdateData
-                );
-
-                expect(updateResponse).toBeDefined();
-                expect(typeof updateResponse.success).toBe('boolean');
-              } catch (updateError) {
-                // Member might not exist, which is expected in test environment
-                expect(updateError).toBeDefined();
-              }
-            }
-          }
-        } catch (error) {
-          expect(error).toBeDefined();
-        }
-      });
-
-      it('should test role hierarchy restrictions', async () => {
-        if (!testTeamId) {
-          // eslint-disable-next-line no-console
-          // eslint-disable-next-line no-console
-          console.log('Skipping test: No test team created');
-          return;
-        }
-
-        // Test that role updates follow hierarchy restrictions
-        // OWNER > ADMIN > EDITOR > VIEWER
-        const testMemberId = 'test-member-id';
-
-        const roleHierarchyTests = [
-          { role: 'OWNER', shouldWork: true }, // Assuming current user has permission
-          { role: 'ADMIN', shouldWork: true },
-          { role: 'EDITOR', shouldWork: true },
-          { role: 'VIEWER', shouldWork: true },
-        ];
-
-        // eslint-disable-next-line no-restricted-syntax
-        for (const roleTest of roleHierarchyTests) {
-          try {
-            const roleUpdateData = { role: roleTest.role };
-            // eslint-disable-next-line no-await-in-loop
-            const updateResponse = await apiClient.put(
-              `/api/v2/team/${testTeamId}/member/${testMemberId}`,
-              roleUpdateData
-            );
-
-            expect(updateResponse).toBeDefined();
-            expect(typeof updateResponse.success).toBe('boolean');
-          } catch (error) {
-            // Member might not exist or permission might be denied
-            expect(error).toBeDefined();
-          }
-        }
-      });
-
-      it('should remove team members with proper validation', async () => {
-        if (!testTeamId) {
-          // eslint-disable-next-line no-console
-          // eslint-disable-next-line no-console
-          console.log('Skipping test: No test team created');
-          return;
-        }
-
-        const testMemberId = 'test-member-id';
-
-        try {
-          const removeResponse = await apiClient.delete(
-            `/api/v2/team/${testTeamId}/member/${testMemberId}`
-          );
-
-          expect(removeResponse).toBeDefined();
-          expect(typeof removeResponse.success).toBe('boolean');
-
-          if (removeResponse.success) {
-            // Should return success confirmation
-            expect(removeResponse.payload || removeResponse.message).toBeDefined();
-          }
-        } catch (error) {
-          // Member might not exist, which is expected in test environment
-          expect(error).toBeDefined();
-        }
-      });
-
-      it('should validate REMOVE_MEMBER permission requirements', async () => {
-        if (!testTeamId) {
-          // eslint-disable-next-line no-console
-          // eslint-disable-next-line no-console
-          console.log('Skipping test: No test team created');
-          return;
-        }
-
-        // Test that member removal requires REMOVE_MEMBER permission
-        const testMemberId = 'test-member-id';
-
-        try {
-          const removeResponse = await apiClient.delete(
-            `/api/v2/team/${testTeamId}/member/${testMemberId}`
-          );
-
-          expect(removeResponse).toBeDefined();
-          expect(typeof removeResponse.success).toBe('boolean');
-
-          // Permission validation should be handled by the API
-          // Either success with proper permissions or failure without them
-        } catch (error) {
-          expect(error).toBeDefined();
-        }
-      });
-
-      it('should enforce business rules for member removal', async () => {
-        if (!testTeamId) {
-          // eslint-disable-next-line no-console
-          // eslint-disable-next-line no-console
-          console.log('Skipping test: No test team created');
-          return;
-        }
-
-        // Test business rules validation
-        // This test validates the API structure for business rule enforcement
-        // rather than testing specific business logic which would require more complex setup
-
-        const testMemberId = 'non-existent-member-id';
-
-        try {
-          const removeResponse = await apiClient.delete(
-            `/api/v2/team/${testTeamId}/member/${testMemberId}`
-          );
-
-          expect(removeResponse).toBeDefined();
-          expect(typeof removeResponse.success).toBe('boolean');
-
-          // When trying to remove non-existent member, should get appropriate response
-          if (removeResponse.success === false) {
-            expect(removeResponse.message).toBeDefined();
-          }
-        } catch (error) {
-          // API call failure is expected for non-existent member
-          expect(error).toBeDefined();
-        }
-      }, 10000);
-    });
-
-    describe('Team Permission Constants Validation', () => {
-      it('should validate team permission constants structure', () => {
-        // Validate the expected permission structure based on TeamPermissionConstants
-        const expectedPermissions = {
-          OWNER: 'All permissions',
-          ADMIN: 'Most permissions except team deletion and bank modifications',
-          EDITOR: 'Main accounting operations',
-          VIEWER: 'Read-only access',
-        };
-
-        expect(expectedPermissions.OWNER).toBeDefined();
-        expect(expectedPermissions.ADMIN).toBeDefined();
-        expect(expectedPermissions.EDITOR).toBeDefined();
-        expect(expectedPermissions.VIEWER).toBeDefined();
-
-        // Validate permission hierarchy
-        const permissionHierarchy = ['OWNER', 'ADMIN', 'EDITOR', 'VIEWER'];
-        expect(permissionHierarchy.length).toBe(4);
-        expect(permissionHierarchy[0]).toBe('OWNER'); // Highest permission
-        expect(permissionHierarchy[3]).toBe('VIEWER'); // Lowest permission
-      });
-
-      it('should validate specific permission operations', () => {
-        // Test the permission operations mentioned in the API documentation
-        const permissionOperations = [
-          'LIST_MEMBER_BY_TEAM_ID',
-          'EDIT_MEMBER_ROLE',
-          'REMOVE_MEMBER',
-          'INVITE_MEMBER',
-        ];
-
-        permissionOperations.forEach((permission) => {
-          expect(typeof permission).toBe('string');
-          expect(permission.length).toBeGreaterThan(0);
-        });
-      });
-    });
-  });
-});
